+++ docker/services/metrics_service.py
@@ -16,6 +16,7 @@ from commons.constants import JOB_STEP_PROCESS_METRICS, META_FILE_NAME, \
     CSV_EXTENSION
 from commons.exception import ExecutorException
 from commons.log_helper import get_logger
+from commons.profiler import profiler
 from models.algorithm import Algorithm
 from models.recommendation_history import RecommendationHistory
 from services.clustering_service import ClusteringService
@@ -274,6 +275,7 @@ class MetricsService:
                 os.remove(file)
         return resulted_files
 
+    @profiler(execution_step=f'instance_clustering')
     def divide_on_periods(self, df, algorithm: Algorithm):
         df = self.divide_by_days(
             df, skip_incomplete_corner_days=True,
@@ -382,16 +384,16 @@ class MetricsService:
         periods = []
         period_start = None
         period_end = None
-        for index, row in df.iterrows():
-            cpu_load = row[COLUMN_CPU_LOAD]
+        for row in df.itertuples():
+            cpu_load = getattr(row, COLUMN_CPU_LOAD)
             if period_start and period_end and np.isnan(cpu_load):
                 periods.append((period_start, period_end))
                 period_start = None
                 period_end = None
             elif not period_start and not np.isnan(cpu_load):
-                period_start = index.time()
+                period_start = row.Index.time()
             elif period_start and not np.isnan(cpu_load):
-                period_end = index.time()
+                period_end = row.Index.time()
         if period_start and period_end:
             periods.append((period_start, period_end))
 
@@ -440,15 +442,15 @@ class MetricsService:
     @staticmethod
     def read_metrics(metric_file_path, algorithm: Algorithm = None,
                      parse_index=True):
-        read_configuration = algorithm.get_read_configuration()
         try:
             if not parse_index:
-                return pd.read_csv(metric_file_path, **read_configuration)
+                return pd.read_csv(metric_file_path,
+                                   **algorithm.read_configuration)
             return pd.read_csv(
                 metric_file_path, parse_dates=True,
                 date_parser=dateparse,
                 index_col=algorithm.timestamp_attribute,
-                **read_configuration)
+                **algorithm.read_configuration)
         except Exception as e:
             _LOG.error(f'Error occurred while reading metrics file: {str(e)}')
             raise ExecutorException(