+++ r8s/r8s_service/adapter_client.py
@@ -561,15 +561,15 @@ class AdapterClient:
                                    payload=request)
 
     def parent_post(self, application_id, description,
-                    clouds, scope, tenant_name=None):
+                    cloud, scope, tenant_name=None):
         request = {
             PARAM_APPLICATION_ID: application_id,
             PARAM_DESCRIPTION: description,
-            PARAM_CLOUDS: clouds,
-            PARAM_SCOPE: scope
+            PARAM_CLOUD: cloud,
+            PARAM_SCOPE: scope,
+            PARAM_TENANT: tenant_name
         }
-        if tenant_name:
-            request[PARAM_TENANT] = tenant_name
+        request = {k: v for k, v in request.items() if v}
 
         return self.__make_request(resource=API_PARENT,
                                    method=HTTP_POST,
@@ -581,34 +581,6 @@ class AdapterClient:
         }
 
         return self.__make_request(resource=API_PARENT,
-                                   method=HTTP_DELETE,
-                                   payload=request)
-
-    def parent_tenant_link_get(self, parent_id):
-        request = {
-            PARAM_PARENT_ID: parent_id
-        }
-
-        return self.__make_request(resource=API_PARENT_TENANT_LINK,
-                                   method=HTTP_GET,
-                                   payload=request)
-
-    def parent_tenant_link_post(self, parent_id, tenant_name):
-        request = {
-            PARAM_PARENT_ID: parent_id,
-            PARAM_TENANT: tenant_name
-        }
-
-        return self.__make_request(resource=API_PARENT_TENANT_LINK,
-                                   method=HTTP_POST,
-                                   payload=request)
-
-    def parent_tenant_link_delete(self, tenant_name):
-        request = {
-            PARAM_TENANT: tenant_name
-        }
-
-        return self.__make_request(resource=API_PARENT_TENANT_LINK,
                                    method=HTTP_DELETE,
                                    payload=request)
 
@@ -666,11 +638,10 @@ class AdapterClient:
                                    method=HTTP_GET,
                                    payload=request)
 
-    def shape_rule_post(self, parent_id, cloud, action, condition, field,
+    def shape_rule_post(self, parent_id, action, condition, field,
                         value, application_id=None):
         request = {
             PARAM_PARENT_ID: parent_id,
-            PARAM_CLOUD: cloud,
             PARAM_RULE_ACTION: action,
             PARAM_CONDITION: condition,
             PARAM_FIELD: field,
@@ -711,10 +682,9 @@ class AdapterClient:
                                    method=HTTP_DELETE,
                                    payload=request)
 
-    def shape_rule_dry_run_get(self, parent_id, cloud):
+    def shape_rule_dry_run_get(self, parent_id):
         request = {
-            PARAM_PARENT_ID: parent_id,
-            PARAM_CLOUD: cloud
+            PARAM_PARENT_ID: parent_id
         }
 
         return self.__make_request(resource=API_SHAPE_RULES_DRY_RUN,